src/requirements.txt
alabaster==0.7.12
apt-clone==0.2.1
apturl==0.5.2
attrs==23.1.0
Babel==2.8.0
beautifulsoup4==4.10.0
blinker==1.6.2
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.1.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
cryptography==3.4.8
cupshelpers==1.0
DateTime==5.1
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.4
distro==1.7.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.17.1
eyeD3==0.8.10
filelock==3.6.0
Flask==2.3.2
ghp-import==2.1.0
grpcio==1.30.2
h11==0.14.0
html2text==2020.1.16
httplib2==0.20.2
idna==3.3
ifaddr==0.1.7
imagesize==1.3.0
IMDbPY==2021.4.18
importlib-metadata==4.6.4
isodate==0.6.1
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.7
MarkupSafe==2.1.2
mergedeep==1.3.4
mkdocs==1.4.3
more-itertools==8.10.0
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.0
onboard==1.4.1
packaging==21.3
PAM==0.4.2
pdoc3==0.10.0
pexpect==4.8.0
Pillow==9.0.1
platformdirs==3.5.1
protobuf==3.12.4
psutil==5.9.0
psycopg2==2.9.2
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pycurl==7.44.1
pyelftools==0.27
Pygments==2.11.2
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyparted==3.11.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-decouple==3.1
python-dotenv==0.19.2
python-gnupg==0.4.8
python-magic==0.4.24
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyyaml_env_tag==0.1
reportlab==3.6.8
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==1.0.0
roman==3.3
SecretStorage==3.3.1
setproctitle==1.2.2
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.1
speedtest-cli==2.1.3
Sphinx==4.3.2
suds==1.1.2
systemd-python==234
texttable==1.6.4
tinycss2==1.1.1
tldextract==3.1.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
Unidecode==1.3.3
urllib3==1.26.5
virtualenv==20.13.0+ds
wadllib==1.3.6
watchdog==3.0.0
webencodings==0.5.1
websocket-client==1.2.3
xdg==5
xkit==0.0.0
xlrd==1.2.0
youtube-dl==2021.12.17
zeep==4.2.1
zipp==1.0.0
zope.interface==6.0
